Finalisation de gestion de map
Le joueur marche de façon fluide et ne marche plus sur les maisons.
Les images sont chargées au début du programme, il n'y a donc quasiment plus de perte de mémoire.

+++ level_editor.h
@@ -7,4 +7,4 @@
 #include "fonctions_sdl.h"
 #include "map.h"
 
-void showEditor(SDL_Renderer *, SDL_Window*, int, int);
+void showEditor(SDL_Renderer *, SDL_Window*, float, float);
